refactor: move ops
